Revert "small update"
This reverts commit 878109a633f3daea8c79ef2e94e4a6d0ac1f3718.

+++ src/main/java/org/onedatashare/server/model/request/UserTransferOptions.java
@@ -36,5 +36,5 @@ public class UserTransferOptions {
     private Integer concurrencyThreadCount;
     private Integer parallelThreadCount;
     private Integer pipeSize;
-    private Long chunkSize;
+    private Integer chunkSize;
 }
